server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Saatwik Vasishtha <saatwik.vasishtha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.114.0"}
sqlalchemy = "^2.0.34"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
langchain = "^0.2.16"
openai = "^1.44.1"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
langchain-mistralai = "^0.1.13"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
langchain-qdrant = "^0.1.3"
langchain-huggingface = "^0.0.3"
matplotlib = "^3.9.2"
pillow = "^11.0.0"
gradio-client = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
